added original compressed code
Some very minor changes were made from the original to the final
version. Specifically, a variable for which person/imageset to run the
model on was modified (in the original code it is by default
Ariel_Sharon, while in the latest version it was changed to
Jennifer_Capriati.

Since there were no apparent quantitative results, but all the results
were qualitative (as far as I could see), this change was only made to
try the model on different persons than only the default.

Additionally, some output files and directories were generated,
specificaly `lfw_prepost.bmp`, `obama.bmp`, and `results/`.

The original zip is included with this commit, the extracted and
modified code in the next commit.